<feature>[ldap]: move LDAP driver related classes to packege ldap.driver
Related: ZSV-5531
