--- plugin/ldap/src/main/java/org/zstack/ldap/LdapMessage.java
@@ -1,7 +1,0 @@
-package org.zstack.ldap;
-
-/**
- */
-public interface LdapMessage {
-    String getEipUuid();
-}
