+++ plugin/ldap/src/main/java/org/zstack/ldap/driver/LdapSearchedResult.java
@@ -1,4 +1,4 @@
-package org.zstack.ldap;
+package org.zstack.ldap.driver;
 
 import java.util.List;
